Bootstrap 4 update + working icon
Site has been upgraded to Bootstrap 4. This involed many changes to the markup, including replacing adding extra rows & changing the text-center class to mx-auto. Also the regular expression in the webpack file was adjusted to allow icon images. A blue and red Djokovic icon is now present in the browser tab.